fix delete area

+++ src/application/areas/area-delete.service.ts
@@ -4,9 +4,19 @@ export class AreaDeleteService {
   constructor(private unitOfWork: IUnitOfWorkInterface) {}
 
   async execute(id: string) {
-    const areaRepository = this.unitOfWork.areaRepository;
+    const { areaRepository, notaRepository } = this.unitOfWork;
     const area = await areaRepository.findOne({ where: { id } });
     if (!area) throw new Error('El regístro no existe!');
-    return areaRepository.delete(area.id);
+    const count = await notaRepository
+      .createQueryBuilder('n')
+      .andWhere(`n.areaId = '${area.id}'`)
+      .getCount();
+    if (count) {
+      await areaRepository.update(area.id, { state: false });
+      return { message: 'Registro desactivado' };
+    } else {
+      await areaRepository.delete(area.id);
+      return { message: 'Registro eliminado!!!' };
+    }
   }
 }
